requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
rembg>=2.0.50
